AG-17862 Move markdownTable/toAbsoluteUrl to ag-website-shared
Hoist the two product-agnostic markdown helpers into the shared subrepo so
the AG-17912 charts port can reuse them, matching the existing shared split
(grid/chartsUrlWithPrefix, parseVersion, getChangelogUrl). Pure relocation:
update the grid importers to @ag-website-shared/markdoc/*, no behaviour change.
