+++ documentation/ag-grid-docs/src/utils/markdoc/renderMarkdocTag.ts
@@ -3,7 +3,9 @@ import featuresData from '@ag-website-shared/components/features-section/DocsFea
 import { FEATURE_MAP } from '@ag-website-shared/components/getting-started/gettingStartedData';
 import { FIGMA_DESIGN_SYSTEM_URL } from '@ag-website-shared/constants';
 import whatsNewData from '@ag-website-shared/content/whats-new/data.json';
+import { markdownTable } from '@ag-website-shared/markdoc/markdownTable';
 import { type MarkdownFramework, fencedCodeBlock } from '@ag-website-shared/markdoc/renderMarkdocToMarkdown';
+import { toAbsoluteUrl } from '@ag-website-shared/markdoc/toAbsoluteUrl';
 import { getDocumentationArchiveUrl } from '@ag-website-shared/utils/getArchiveUrl';
 import { getChangelogUrl } from '@ag-website-shared/utils/getChangelogUrl';
 import { parseVersion } from '@ag-website-shared/utils/parseVersion';
@@ -17,11 +19,9 @@ import { readFileSync } from 'node:fs';
 import path from 'node:path';
 
 import gridSeedProjects from '../../content/seed-projects/grid-seed-projects.json';
-import { markdownTable } from './markdownTable';
 import { type VersionEntry, buildMajorTable } from './renderMajorTable';
 import { buildMatrixTable } from './renderMatrixTable';
 import { type ModuleNode, buildModuleMappingsTable } from './renderModuleMappings';
-import { toAbsoluteUrl } from './toAbsoluteUrl';
 
 interface RenderMarkdocTagParams {
     tag: string;
